@@ -1,61 +1,23 @@
-import React from 'react';
+import React, {useEffect, useState} from 'react';
 import PropTypes from 'prop-types';
 import SocketIO from 'socket.io-client';
 import {SocketContext} from './SocketContext';
 import {warning, debug} from './utils';
 
-// eslint-disable-next-line require-jsdoc
-class Socket extends React.Component {
-  static contextType = SocketContext;
+/**
+ * Socket - the socket.io Socket component.
+ * @param {object} props
+ * @return {JSX.Element}
+ */
+const Socket = (props) => {
+  const [socket, setSocket] = useState(null);
 
-  // eslint-disable-next-line require-jsdoc
-  constructor(props) {
-    super(props);
-
-    // eslint-disable-next-line new-cap
-    this.socket = SocketIO(props.uri, this.mergeOptions(props.options));
-
-    this.socket.status = 'initialized';
-
-    this.socket.on('connect', () => {
-      console.info('[Socket] connect');
-      this.socket.status = 'connected';
-      debug('connected');
-    });
-
-    this.socket.on('disconnect', () => {
-      console.info('[Socket] disconnect');
-      this.socket.status = 'disconnected';
-      debug('disconnect');
-    });
-
-    this.socket.on('error', (err) => {
-      this.socket.status = 'failed';
-      warning('error', err);
-    });
-
-    this.socket.on('reconnect', (data) => {
-      this.socket.status = 'connected';
-      debug('reconnect', data);
-    });
-
-    this.socket.on('reconnect_attempt', () => {
-      debug('reconnect_attempt');
-    });
-
-    this.socket.on('reconnecting', () => {
-      this.socket.status = 'reconnecting';
-      debug('reconnecting');
-    });
-
-    this.socket.on('reconnect_failed', (error) => {
-      this.socket.status = 'failed';
-      warning('reconnect_failed', error);
-    });
-  }
-
-  // eslint-disable-next-line require-jsdoc
-  mergeOptions(options = {}) {
+  /**
+   * mergeOptions - merge socket.io options.
+   * @param {object} options - the options to merge
+   * @return {object}
+   */
+  const mergeOptions = (options = {}) => {
     const defaultOptions = {
       reconnection: true,
       reconnectionAttempts: Infinity,
@@ -66,18 +28,66 @@ class Socket extends React.Component {
       rejectUnauthorized: true,
     };
     return {...defaultOptions, ...options};
-  }
+  };
 
-  // eslint-disable-next-line require-jsdoc
-  render() {
-    return (
-      <SocketContext.Provider value={this.socket}>
-        {React.Children.only(this.props.children)}
-      </SocketContext.Provider>
-    );
-  }
-}
+  /**
+   * Similar to componentDidMount and componentDidUpdate.
+   */
+  useEffect(() => {
+    if (!socket) {
+      // eslint-disable-next-line new-cap
+      setSocket(SocketIO(props.uri, mergeOptions(props.options)));
+    } else {
+      socket.status = 'initialized';
 
+      socket.on('connect', () => {
+        console.info('[Socket] connect');
+        socket.status = 'connected';
+        debug('connected');
+      });
+
+      socket.on('disconnect', () => {
+        console.info('[Socket] disconnect');
+        socket.status = 'disconnected';
+        debug('disconnect');
+      });
+
+      socket.on('error', (err) => {
+        socket.status = 'failed';
+        warning('error', err);
+      });
+
+      socket.on('reconnect', (data) => {
+        socket.status = 'connected';
+        debug('reconnect', data);
+      });
+
+      socket.on('reconnect_attempt', () => {
+        debug('reconnect_attempt');
+      });
+
+      socket.on('reconnecting', () => {
+        socket.status = 'reconnecting';
+        debug('reconnecting');
+      });
+
+      socket.on('reconnect_failed', (error) => {
+        socket.status = 'failed';
+        warning('reconnect_failed', error);
+      });
+    }
+  }, [socket]);
+
+  /**
+   * Renders the React component.
+   * @return {JSX.Element} - React markup for component.
+   */
+  return (
+    <SocketContext.Provider value={socket}>
+      {React.Children.only(props.children)}
+    </SocketContext.Provider>
+  );
+};
 Socket.propTypes = {
   options: PropTypes.object.isRequired,
   uri: PropTypes.string.isRequired,
