@@ -5,8 +5,62 @@ from python.libs import TSV
 from mne_bids import write_raw_bids, BIDSPath
 
 
+class Anonymize:
+    file_path = ''
+    header = []
+
+    def __init__(self, data):
+        self.file_path = data['file_path']
+
+        # read EDF file from file_path,
+        file_in = EDF.EDFReader(fname=self.file_path)
+
+        # read header of EDF file.
+        self.header = file_in.readHeader()
+        file_in.close()
+
+        # print('header is ')
+        # print(self.header)
+        # print('subject_id:')
+        # print(self.header[0]['subject_id'])
+        # print('recording_id:')
+        # print(self.header[0]['recording_id'])
+        # print('day:')
+        # print(self.header[0]['day'])
+        # print('month:')
+        # print(self.header[0]['month'])
+        # print('year:')
+        # print(self.header[0]['year'])
+        # print('hour:')
+        # print(self.header[0]['hour'])
+        # print('minute:')
+        # print(self.header[0]['minute'])
+        # print('second:')
+        # print(self.header[0]['second'])
+
+    def get_header(self):
+        return self.header
+
+    def set_header(self, key, value):
+        self.header[0][key] = value
+
+    def make_copy(self, new_file):
+        header = self.get_header()
+        file_in = EDF.EDFReader(fname=self.file_path)
+        file_out = EDF.EDFWriter()
+        file_out.open(new_file)
+        file_out.writeHeader(header)
+        meas_info = header[0]
+        for i in range(meas_info['n_records']):
+            data = file_in.readBlock(i)
+            file_out.writeBlock(data)
+        file_in.close()
+        file_out.close()
+
+
 class Converter:
     m_info = ''
+
     # data: {
     #  file_path: '', // where file located.
     #  bids_directory: '', // where to output.
@@ -17,6 +71,8 @@ class Converter:
         self.to_bids(
             file=data['file_path'],
             bids_directory=data['bids_directory'],
+            subject_id=data['subject_id'],
+            output_time=data['output_time'],
             read_only=data['read_only'],
             line_freq=data['line_freq']
         )
@@ -36,6 +92,8 @@ class Converter:
     def to_bids(self,
                 file,
                 bids_directory,
+                subject_id,
+                output_time,
                 task='test',
                 ch_type='seeg',
                 read_only=False,
@@ -58,8 +116,12 @@ class Converter:
             raw.set_channel_types({ch: ch_type for ch in raw.ch_names})
             print('VIEW 2:')
             print(raw)
+            os.makedirs(bids_directory + os.path.sep + output_time, exist_ok=True)
+            bids_directory = bids_directory + os.path.sep + output_time
             bids_root = bids_directory
+            m_info['subject_id'] = subject_id  # 'alizee'
             subject = m_info['subject_id'].replace('_', '').replace('-', '').replace(' ', '')
+            # subject = subject_id  # 'alizee
             print('LOOK:')
             print(subject)
             # subject = 'alizee'  # modified will output sub-alizee
@@ -81,10 +143,19 @@ class Converter:
                 'preload': False,
                 'verbose': None
             }
-            write_raw_bids(raw, bids_basename, overwrite=False, verbose=False)
+            write_raw_bids(raw, bids_basename, anonymize=dict(daysback=33630), overwrite=False, verbose=False)
             print('finished')
         else:
             print('File not found or is not file: %s', file)
+
+
+class Time:
+    def __init__(self):
+        print('- Time: init started.')
+        # print(data)
+        from datetime import datetime
+        now = datetime.now()
+        self.latest_output = now.strftime("%Y-%m-%d-%Hh%Mm%Ss")
 
 
 class Modifier:
