@@ -9,15 +9,24 @@ import '../css/Welcome.css';
  */
 const Welcome = (props) => {
   /**
+   * openGitHub - Navigate browser to pycat.
+   */
+  const openGitHub = () => {
+    const myAPI = window['myAPI'];
+    myAPI.visitGitHub();
+  };
+
+  /**
    * Renders the React component.
    * @return {JSX.Element} - React markup for component.
    */
   return props.visible ? (
     <>
-      <div className={'title'}>
+      <span className={'title'}>
         Welcome to <b>pyCat!</b>
-      </div>
+      </span>
       <div className={'info'}>
+        <br/><br/><br/>
         <p className={'font-large'}>Hello user,</p>
         <p className={'font-medium'}><b>pyCat</b>&nbsp;
           is a simple tool for de-identification of iEEG datasets.
@@ -29,27 +38,33 @@ const Welcome = (props) => {
           <b>The Configuration tab allows to:</b>
         </p>
         <ul>
-          <li>todo ...</li>
-          <li>todo ...</li>
-          <li>todo ...</li>
+          <li>Select the file.edf, events.tsv and BIDS output directory.</li>
+          <li>Set the line_freq and LORIS metadata.</li>
+          <li>Anonymize the iEEG header data.</li>
         </ul>
         <p>
           <b>The iEEG to BIDS tab allows to:</b>
         </p>
         <ul>
-          <li>todo ...</li>
-          <li>todo ...</li>
+          <li>Review your configurations.</li>
+          <li>Review your LORIS metadata.</li>
+          <li>Review your iEEG header data.</li>
         </ul>
         <p>
           <b>The Validator tab allows to:</b>
         </p>
         <ul>
-          <li>todo ...</li>
+          <li>Confirms the BIDS structure is valid.</li>
         </ul>
         <p className={'font-medium'}>
           You may begin your task by following the menu above.
           Please remember to always backup your data!
         </p>
+      </div>
+      <div className={'footer'}>
+        Powered by <a className={'open-source'} onClick={openGitHub}>
+        open source software</a>.<br/>
+        Copyright © 2021 MCIN.
       </div>
     </>
   ) : null;
